plan: export `MockTable` and `MockContext` to be used in cascades package
